+++ gerenciaBanco/src/main/java/com/mycompany/gerenciabanco/User.java
@@ -1,38 +1,33 @@
 package com.mycompany.gerenciabanco;
 
 import java.util.*;
-import java.io.*;
 
 public class User {
     
-    public char name; // Declara as variaveis
-    public char lastName;
-    public long document;
-    public int  flagbd;
+    public String name; // Declara as variaveis
+    public String lastName;
+    public String document;
 
     public User(){} // Metodo construtor
 
-    public long coletaDados(){
+    public String[] coletaDados(){
         Scanner myObj = new Scanner(System.in);  // Scanner para ler inputs
-        System.out.println("Digite o seu documento (apenas numeros)");
 
-        document = myObj.nextLong();  // Leitura do input do usuario
-        System.out.println("Voce digitou o documento: " + document);
-        return document;
+        System.out.println("Digite o seu documento (apenas numeros)");
+        document = myObj.nextLine();  // Leitura do input do usuario
+        /*System.out.println("Voce digitou o documento: " + document + "\n");*/
+        
+        System.out.println("Digite o seu nome (minusculo)");
+        name = myObj.nextLine();  // Leitura do input do usuario
+        /*System.out.println("Voce digitou o nome: " + name + "\n");*/
+
+        System.out.println("Digite o seu sobrenome (minusculo)");
+        lastName = myObj.nextLine();  // Leitura do input do usuario
+        /*System.out.println("Voce digitou o sobrenome: " + lastName + "\n");*/
+        
+        String[] tempString = {document, name, lastName}; 
+        return tempString;
     }
-    
-    public void checaBd(){
-        try{
-            File f = new File("database.txt"); // Indica qual arquivo queremos manipular
-            if(f.createNewFile()){ //Tenta criar um arquivo com o nome dado
-                System.out.println("Database criado!"); //Se nao der erro ele cria
-            }else{
-                System.out.println("Database ja existe!"); //Se arquivo ja existir ele avisa
-            }
-        }catch(IOException e){
-            System.out.println("Ocorreu um erro:");
-            e.printStackTrace(); // Caso falhe, retorna o erro
-        }   
-    }
+
 
 }
